+++ akshare/bond/bond_investing.py
@@ -8,10 +8,11 @@ https://cn.investing.com/rates-bonds/
 import re
 
 import pandas as pd
-import requests
 from bs4 import BeautifulSoup
 
 from akshare.index.cons import short_headers, long_headers
+
+from akshare.utils.ak_session import session
 
 
 def _get_global_country_name_url() -> dict:
@@ -22,7 +23,7 @@ def _get_global_country_name_url() -> dict:
     :rtype: dict
     """
     url = "https://cn.investing.com/rates-bonds/"
-    res = requests.get(url, headers=short_headers, timeout=30)
+    res = session.get(url, headers=short_headers, timeout=30)
     soup = BeautifulSoup(res.text, "lxml")
     name_url_option_list = soup.find("select", attrs={"name": "country"}).find_all("option")[1:]
     url_list = [item["value"] for item in name_url_option_list]
@@ -41,7 +42,7 @@ def bond_investing_global_country_name_url(country: str = "中国") -> dict:
     """
     name_url_dict = _get_global_country_name_url()
     url = f"https://cn.investing.com{name_url_dict[country]}"
-    res = requests.get(url, headers=short_headers, timeout=30)
+    res = session.get(url, headers=short_headers, timeout=30)
     soup = BeautifulSoup(res.text, "lxml")
     url_list = [
         item.find("a")["href"] for item in soup.find_all(attrs={"class": "plusIconTd"})
@@ -83,7 +84,7 @@ def bond_investing_global(
     period_map = {"每日": "Daily", "每周": "Weekly", "每月": "Monthly"}
     name_code_dict = bond_investing_global_country_name_url(country)
     temp_url = f"https://cn.investing.com/{name_code_dict[index_name]}-historical-data"
-    res = requests.get(temp_url, headers=short_headers, timeout=30)
+    res = session.get(temp_url, headers=short_headers, timeout=30)
     soup = BeautifulSoup(res.text, "lxml")
     title = soup.find("h2", attrs={"class": "float_lang_base_1"}).get_text()
     data = soup.find_all(text=re.compile("window.histDataExcessInfo"))[0].strip()
@@ -100,7 +101,7 @@ def bond_investing_global(
         "action": "historical_data",
     }
     url = "https://cn.investing.com/instruments/HistoricalDataAjax"
-    res = requests.post(url, data=payload, headers=long_headers, timeout=60)
+    res = session.post(url, data=payload, headers=long_headers, timeout=60)
     df_data = pd.read_html(res.text)[0]
     df_data.columns = [
         '日期',
@@ -130,6 +131,6 @@ if __name__ == "__main__":
         index_name="中国10年期国债",
         period="每日",
         start_date="20100101",
-        end_date="20210314",
+        end_date="20210318",
     )
     print(bond_investing_global_df)