+++ akshare/index/index_baidu.py
@@ -5,11 +5,8 @@ Date: 2020/9/28 13:53
 Desc: 百度指数
 感谢 https://cloudcrawler.club/categories/2019%E5%B9%B4%E6%9C%AB%E9%80%86%E5%90%91%E5%A4%8D%E4%B9%A0/
 """
-import matplotlib.pyplot as plt
 import pandas as pd
 import requests
-
-plt.rcParams["font.sans-serif"] = ["SimHei"]  # 显示中文标签
 
 
 def decrypt(t: str, e: str) -> str:
@@ -210,20 +207,13 @@ if __name__ == "__main__":
         word="python", start_date="2020-01-01", end_date="2020-09-14", cookie=cookie
     )
     print(data)
-    data.dropna(inplace=True)
-    data.plot()
-    plt.show()
+
     data = baidu_info_index(
         word="口罩", start_date="2019-07-03", end_date="2020-09-21", cookie=cookie
     )
     print(data)
-    data.dropna(inplace=True)
-    data.plot()
-    plt.show()
+
     data = baidu_media_index(
         word="金融科技", start_date="2020-01-01", end_date="2020-09-20", cookie=cookie
     )
     print(data)
-    data.dropna(inplace=True)
-    data.plot()
-    plt.show()